+++ src/react/data/MutationData.ts
@@ -3,7 +3,7 @@ import { equal as isEqual } from '@wry/equality';
 import { DocumentType } from '../parser/parser';
 import { ApolloError } from '../../errors/ApolloError';
 import {
-  MutationOptions,
+  MutationDataOptions,
   MutationTuple,
   MutationFunctionOptions,
   MutationResult
@@ -27,7 +27,7 @@ export class MutationData<
     result,
     setResult
   }: {
-    options: MutationOptions<TData, TVariables>;
+    options: MutationDataOptions<TData, TVariables>;
     context: any;
     result: MutationResult<TData>;
     setResult: (result: MutationResult<TData>) => any;